Implemented HFile backup to archive
Essentially, whenever an hfile would be deleted, it is instead
moved to the archive directory. In this impl, the
archive directory is on a per table basis, but defaults to '.archive'.
Removing hfiles occurs in three places - compaction, merge and catalog janitor.
The former and two latter are distinctly different code paths, but but did
pull out some similarities. The latter two end up calling the same method,
so there should be a reasonable amount of overlap.

Updated the HMasterInterface to pass the calls onto the zookeeper.
Added a zk listener to handle updates from the master to the RS to backup.
Added a utility for removing files and finding archive directories
Added tests for the regionserver and catalogjanitor region deletion.
Added creation of manager in regionserver.
Updating as per Ted's comments 4/5/12
Add sychronous solution that waits for all RS to agree

Concurent patches:
- ZKUtil.deleteRecursively needds to be clean
- ZKUtil.deleteRecursively needs to be recoverable

Recent:
Fixing bugs from branch -r2.
Not sure what went wrong, but something did.
Also, adding fixes from Lars and Ted

Even more recent:
Fixed enable/disable archving test
Going back to using an archive tracker.
Need to abstract out as a whole object again