Got snapshots working.
Lots of work to get the orchestration working properly, on the RS side.
Also, added use of countdown latches for the snapshot sentinel (ready for multi-threading).
Also, did some general cleanup (licenes, moving things out of HConstants).
Smaller bug fixes for the master and RS sides of the impl
Impl of listSnapshots and deleteSnapshots as well.