Impl of TimePartitonedStore.
A store that wraps an existing Store and routes writes to a snapshot or the primary kvset.
Splitting is based on a 'split point' timestamp. Less than the TS goes to the snapshot and greater
go to the primary.

This ensures that the snapshot has all writes up to the the given timestamp.
Caveat: if a write exists with a timestamp in the past, it will be put into the snapshot, rather
than into the primary kvset. However, this is fine - at worst we have some extra data in the snapshot,
but never miss data that should be in the kvset.