Implementing basic cross-process error handling.
This is a pretty simple paradigm that has been generalized for multiple processes.
Each 'task' monitored has multiple 'parts', each part potentially running in its own thread.
Each task has a single ErrorMonitor that is shared between all processes (BaseBoundErrorMonitor helps impl this).
Each task can then check the error monitor for errors to the overal task or notify the task of errors.
Each task will probably have their own type of errors and listeners, rather than dealing with an array of objects as
error information. This manipulation is handled by the ErrorListenerBridge.
Further, we may want to ignore errors if they don't meet basic information - this is handled by the ReceiveErrorChecker.
If the task is run multiple times, a generic factory (subclassable) is provided to make it easy to make an ErrorMonitor
that converts between types as described above.
Futher, the factory also lets you inject errors into the runner via a FaultInjector, but with minimal overhead to running
tasks using a non-injecting error monitor.