Start of the impl of snapshot error handling on the abstract framework.
Not yet integrated into the current snapshot stuff, but seems basically there.
Old error handling still in place (needs to be removed).
On the upside, it ends up being really clean and simple to implement.

Starting to think the ReceiverErrorChecker is a little excessive and can be handled in the listenerBridge,
but let's see how it plays out in the final implementation: