Logs and hfile cleaning
Switching to using a single SnapshotFileCache
This is more of a composition based tooling, which means a much
simpler architecture and can be more easily monitored.

The whole CleanerChoreUtil thing goes away and all the inheritance stuff.
Now each cleaner (hfile and log) has its own cache that it can reference and use to keep track of its files.

TODO:
 - fix directory saving when loading the cache