Reimplementing error handling on top of javax.management's Notification mechanism.
This takes care of needing a special timer, tracking listeners, passing around
extra info, and binding to types. Also, updating the offline snapshot implementation
to the new error handling - almost no major changes (real test is the distributed, but
that should be just as smooth).

Reevaluated what the actual requirements would need to be for an error handling framework
after looking poking around a couple notification implementations. javax.management's
utils do everything we are doing, but in an already established way. The only thing it
doesn't do is weak referencing listeners to allow non-explicit (lazy) removal - that was a
pretty trivial add.

I did end up sacrificing the FaultInjection tooling, but this:
 1) makes it easier to understand the code
 2) can still be implemented via Mockito.doAnswer() and checking the current stack.

One of the main difficulties in using the notification framework is how to deal with:
 1) standard errors (e.g. server abort, timeout)
 2) easily adding errors for a use-case.

The intended use case would be that you have one NotificationBroadcaster that listens for
high level failures (e.g. aborts, non-recoverable errors, etc) and then a per-task instance
ExceptionSnare that listens for errors from the broadcaster and also for instance errors, of which
there could be many different exceptions.

This all gets wrapped into the ExceptionSnare (which like the previous ErrorSnare) still tracks the
first received error, with the option to have your own snare that maps from a generic exception to an
instance of an exception (inspired by guava's CheckedFuture).

I don't really like how the exception tagging currently works - that will take some more thinking to
get right, but it seems like using standard tags and per-task tags combined with some sort of filtering
would be the right path.